requirements.txt
pyroblack
tgcrypto
uvicorn
fastapi[all]
python-dotenv
aiofiles==23.2.1
aiohttp==3.9.5
curl_cffi
techzdl>=1.2.6
Pillow>=9.5.0
python-multipart
ffmpeg-python